- Code tidy in lib folder c code in response to a splint run.   removed some unreferenced variables and corrected some dodgy   casts between variables. No changed noticed in X11 or GTK   preformance on OSX (64bit GTK on OSX 10.6 is still problematic).
  This commit to allow testing on other platforms

Testing
- Compile OSX GCC 4.6 X11 and GTK
- splint runs with -weak argument
- Interactive tests of prj on X11 and GTK. Noticed change
  of focus to single room or use of rotate view pulldown
  option still corrupts wireframe view. Changes did no
  damage but did not correct issue.




Former-commit-id: 89ed13c92671052e2b8918279d04e29b9db83619

+++ src/lib/esp_ask.c
@@ -569,7 +569,7 @@ void askdialog248_(char *q1, char *reply,long int *ier, int lenq1, int lenrep)
          lnq1 = 0;	/* detect the actual size of reply_local */
          f_to_c_l(reply_local,&lenrep,&lnq1);
 /* debug g_print ("askdialog248 local %d is %s",lnq1,reply_local); */
-         g_snprintf (reply, lnq1+1, "%s", reply_local); /* copy reply_local into reply +1 for null */
+         g_snprintf (reply, (gulong)lnq1+1, "%s", reply_local); /* copy reply_local into reply +1 for null */
 /* debug g_print ("askdialog248 reply is %s",reply); */
          g_free (reply_local);
          no_valid_event = FALSE;
